Clean up output for MiddlewareInterface._get_all_pipeline_files.
The output was annotated as an iterable, specced as a sequence with
"undefined order" (which defeats the main benefit of using a sequence),
and implemented as a list with possible duplicates. I've changed the
docs to self-consistently say that it's a collection of arbitrary type,
and implemented as a set internally (though the spec still does not
guarantee deduplication).